refactor: move last-used login cookie to frontend, trim public API
The last-login cookie is written and read entirely client-side, so the
contract belongs in the frontend, not @lightdash/common. Move it to
features/users/utils/lastLoginMethod and expose only readLastLoginMethod /
writeLastLoginMethod (+ the type) — the encode/decode/cookie-name internals are
no longer exported. Record the authenticated user's email on password login
rather than the form value, and exercise the public read/write API in the test.
